✅ 完成同步节点

+++ synchronizer/core/models/config_models.go
@@ -31,5 +31,5 @@ type Mq struct {
 	Kafka Kafka `yaml:"kafka"`
 }
 type Snowflake struct {
-	MachineID int `yaml:"machineID"`
+	MachineID int64 `yaml:"machineID"`
 }
